Add a singleton and use it to lload my json instead of loading json in the App Delegate
