+++ Artroom1/AppDelegate.swift
@@ -12,23 +12,12 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
-
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-      
-        func loadJson(_ fileName: String) -> [Artwork]? {
-        if let path = Bundle.main.path(forResource: "data", ofType: "json") {
-            do {
-                let data = try Data(contentsOf: URL(fileURLWithPath: path))
-                let decoder = JSONDecoder()
-                let jsonData = try decoder.decode(ResponseData.self, from: data)
-                return jsonData.artwork
-            } catch {
-                print("error:\(error)")
-            }
-            }
-            return nil
-        }
-        arrayOfArtworks = loadJson("data")!
+        
+        //        Originally I tried to do this here:
+        //        ArtworksDatabase.shared.apply()
+        //        but it didn't do anything. 
         
         return true
     }
