--- Artroom1/ArtPieceDB.swift
@@ -1,11 +1,0 @@
-//
-//  ArtPieceDB.swift
-//  Artroom1
-//
-//  Created by Taylor Lindsay on 10/31/18.
-//  Copyright © 2018 Taylor Lindsay at Babbel. All rights reserved.
-//
-
-import Foundation
-
-var arrayOfArtworks: [Artwork] = []
